# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PETUK Restaurant Business Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listAdminUsers
      tags: [users]
      summary: List admin users
      responses:
        "200":
          description: Admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [users]
      summary: Create admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /users/{id}:
    put:
      operationId: updateAdminUser
      tags: [users]
      summary: Update admin user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteAdminUser
      tags: [users]
      summary: Delete admin user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── SETTINGS ──────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get restaurant settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, username, displayName, role, station, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        station:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    AdminUserInput:
      type: object
      required: [username, displayName, role, station]
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
        role:
          type: string
        station:
          type: string
        isActive:
          type: boolean

    # ── MENU ──────────────────────────────────────────────────────────────
...
    Settings:
      type: object
      required: [id, restaurantName, address, phone, taxEnabled, taxRate, taxName, currency, deliveryFee, deliveryFeeEnabled, pointsPer100Taka, pointsRedemptionRate, maxCashierDiscountPercent, maxCashierDiscountAmount]
      properties:
        id:
          type: integer
        restaurantName:
          type: string
        address:
          type: string
        phone:
          type: string
        taxEnabled:
          type: boolean
        taxRate:
          type: number
        taxName:
          type: string
        currency:
          type: string
        paperWidth:
          type: integer
        deliveryFee:
          type: number
        deliveryFeeEnabled:
          type: boolean
        pointsPer100Taka:
          type: integer
        pointsRedemptionRate:
          type: number
        maxCashierDiscountPercent:
          type: number
        maxCashierDiscountAmount:
          type: number
        syncApiKey:
          type: string
    SettingsUpdate:
      type: object
      properties:
        restaurantName:
          type: string
        address:
          type: string
        phone:
          type: string
        taxEnabled:
          type: boolean
        taxRate:
          type: number
        taxName:
          type: string
        currency:
          type: string
        paperWidth:
          type: integer
        deliveryFee:
          type: number
        deliveryFeeEnabled:
          type: boolean
        pointsPer100Taka:
          type: integer
        pointsRedemptionRate:
          type: number
        maxCashierDiscountPercent:
          type: number
        maxCashierDiscountAmount:
          type: number
        syncApiKey:
          type: string

    # ── DASHBOARD ─────────────────────────────────────────────────────────
...